+++ src/OpenTelemetry/Sdk.cs
@@ -14,9 +14,13 @@
 // limitations under the License.
 // </copyright>
 
+#nullable enable
+
+using System;
 using System.Diagnostics;
 using OpenTelemetry.Context.Propagation;
 using OpenTelemetry.Internal;
+using OpenTelemetry.Logs;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
 
@@ -51,7 +55,39 @@ namespace OpenTelemetry
         /// <param name="textMapPropagator">TextMapPropagator to be set as default.</param>
         public static void SetDefaultTextMapPropagator(TextMapPropagator textMapPropagator)
         {
+            Guard.ThrowIfNull(textMapPropagator);
+
             Propagators.DefaultTextMapPropagator = textMapPropagator;
+        }
+
+        /// <summary>
+        /// Creates a <see cref="OpenTelemetryLoggerOptions"/> which is used to build
+        /// an <see cref="OpenTelemetryLoggerProvider"/>. In a typical application, a single
+        /// <see cref="OpenTelemetryLoggerProvider"/> is created at application startup and disposed
+        /// at application shutdown. It is important to ensure that the provider is not
+        /// disposed too early.
+        /// </summary>
+        /// <returns><see cref="OpenTelemetryLoggerOptions"/> instance, which is used to build a <see cref="OpenTelemetryLoggerProvider"/>.</returns>
+        public static OpenTelemetryLoggerOptions CreateLoggerProviderBuilder()
+        {
+            return new OpenTelemetryLoggerOptionsSdk(configure: null);
+        }
+
+        /// <summary>
+        /// Creates a <see cref="OpenTelemetryLoggerOptions"/> which is used to build
+        /// an <see cref="OpenTelemetryLoggerProvider"/>. In a typical application, a single
+        /// <see cref="OpenTelemetryLoggerProvider"/> is created at application startup and disposed
+        /// at application shutdown. It is important to ensure that the provider is not
+        /// disposed too early.
+        /// </summary>
+        /// <param name="configure">Configuration action.</param>
+        /// <returns><see cref="OpenTelemetryLoggerOptions"/> instance, which is used to build a <see cref="OpenTelemetryLoggerProvider"/>.</returns>
+        public static OpenTelemetryLoggerOptions CreateLoggerProviderBuilder(
+            Action<OpenTelemetryLoggerOptions> configure)
+        {
+            Guard.ThrowIfNull(configure);
+
+            return new OpenTelemetryLoggerOptionsSdk(configure);
         }
 
         /// <summary>