+++ src/OpenTelemetry/CHANGELOG.md
@@ -35,8 +35,8 @@
   `OpenTelemetryLoggerProvider` directly
   ([#3489](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3489))
 
-* Added support for dependency injection scenarios when configuring
-  `OpenTelemetryLoggerProvider`
+* Added `Sdk.CreateLoggerProviderBuilder` method and support for dependency
+  injection scenarios when configuring `OpenTelemetryLoggerProvider`
   ([#3504](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3504))
 
 ## 1.3.0